feat: add delete button in PostPage, change to const to new types for better order bugs: fix newPost and newComment state.

+++ src/Post/PostItem.js
@@ -12,8 +12,10 @@ class PostItem extends Component {
     }
 
     deletePost = (id) => {
-        const {deletePost} = this.props
-        deletePost(id)
+        if (window.confirm("Are you sure you want to delete this post?")) {
+            const {deletePost} = this.props
+            deletePost(id)
+        }
     }
 
     editPost = (id) => {
